+++ api/models/User.js
@@ -1,18 +1,18 @@
 const mongoose = require('mongoose'),
-    Schema = mongoose.Schema;
+  Schema = mongoose.Schema;
 
 // Storing basic user metadata for a user
 const userSchema = Schema(
-    {
-        name: { type: String, default: '' },
-        password: { type: String, default: '' },
-        dateCreated: { type: Date, default: new Date() },
-        lastUsed: { type: Date, default: new Date() }
-    },
+  {
+    name: { type: String, default: '' },
+    password: { type: String, default: '' },
+    dateCreated: { type: Date, default: new Date() },
+    lastUsed: { type: Date, default: new Date() }
+  },
 
-    {
-        collection: 'usersDB'
-    }
+  {
+    collection: 'usersDB'
+  }
 );
 
 // Allowing efficient querying by setting what fields to query using
