+++ app.js
@@ -1,10 +1,10 @@
 'use strict';
 
-var SwaggerExpress = require('swagger-express-mw');
-var app = require('express')();
+const SwaggerExpress = require('swagger-express-mw');
+const app = require('express')();
 module.exports = app; // for testing
 
-var config = {
+const config = {
   appRoot: __dirname // required config
 };
 
@@ -14,18 +14,15 @@ SwaggerExpress.create(config, function(err, swaggerExpress) {
   // install middleware
   swaggerExpress.register(app);
 
-  var port = process.env.PORT || 10010;
+  const port = process.env.PORT || 4000;
   app.listen(port);
-
-  if (swaggerExpress.runner.swagger.paths['/authentication']) {
-    console.log('try this:\ncurl http://127.0.0.1:' + port + '/authentication?name=Scott');
-  }
 });
 
 // Using Mongoose to blackbox the MongoDB connectivity
 const secretURI = require('./mongodb-connectivity-uri');
 const mongoose = require('mongoose');
-mongoose.connect(secretURI, { useNewUrlParser: true });
+
+mongoose.connect(secretURI, { useNewUrlParser: true, useUnifiedTopology: true, useFindAndModify: false });
 
 const db = mongoose.connection;
 db.on('error', (err) => console.log(`Connection error:\n${err}`));
