+++ package.json
@@ -2,7 +2,7 @@
   "name": "seraas-user-management-service",
   "version": "0.0.1",
   "private": true,
-  "description": "New Swagger API Project",
+  "description": "User Management Service for Speech Emotion Recognition as a Service",
   "keywords": [],
   "author": "",
   "license": "",
